mysql module depend on rdb
