+++ mysql/mysql-side/mysql-async-side/pom.xml
@@ -16,27 +16,16 @@
 
     <packaging>jar</packaging>
 
+    <properties>
+        <sql.side.mysql.core.version>1.0-SNAPSHOT</sql.side.mysql.core.version>
+    </properties>
+
     <dependencies>
-        <!--异步数据库连接-->
-        <!--需要提取到对应的插件-->
-        <dependency>
-            <groupId>io.vertx</groupId>
-            <artifactId>vertx-jdbc-client</artifactId>
-            <version>3.5.2</version>
-        </dependency>
-
-        <dependency>
-            <groupId>io.vertx</groupId>
-            <artifactId>vertx-core</artifactId>
-            <version>3.5.2</version>
-        </dependency>
-
         <dependency>
             <groupId>com.dtstack.flink</groupId>
             <artifactId>sql.side.mysql.core</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <version>${sql.side.mysql.core.version}</version>
         </dependency>
-
     </dependencies>
 
     <build>
@@ -87,12 +76,12 @@
                             <tasks>
                                 <copy todir="${basedir}/../../../plugins/mysqlasyncside">
                                     <fileset dir="target/">
-                                        <include name="${project.artifactId}-${project.version}.jar" />
+                                        <include name="${project.artifactId}-${project.version}.jar"/>
                                     </fileset>
                                 </copy>
 
                                 <move file="${basedir}/../../../plugins/mysqlasyncside/${project.artifactId}-${project.version}.jar"
-                                      tofile="${basedir}/../../../plugins/mysqlasyncside/${project.name}.jar" />
+                                      tofile="${basedir}/../../../plugins/mysqlasyncside/${project.name}.jar"/>
                             </tasks>
                         </configuration>
                     </execution>
