+++ mysql/mysql-side/mysql-side-core/src/main/java/com/dtstack/flink/sql/side/mysql/table/MysqlSideParser.java
@@ -16,40 +16,30 @@
  * limitations under the License.
  */
 
- 
 
 package com.dtstack.flink.sql.side.mysql.table;
 
-import com.dtstack.flink.sql.table.AbsSideTableParser;
+import com.dtstack.flink.sql.side.rdb.table.RdbSideParser;
 import com.dtstack.flink.sql.table.TableInfo;
-import com.dtstack.flink.sql.util.MathUtil;
 
 import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Reason:
  * Date: 2018/7/25
  * Company: www.dtstack.com
+ *
  * @author xuchao
  */
 
-public class MysqlSideParser extends AbsSideTableParser {
+public class MysqlSideParser extends RdbSideParser {
+
+    private static final String CURR_TYPE = "mysql";
 
     @Override
     public TableInfo getTableInfo(String tableName, String fieldsInfo, Map<String, Object> props) {
-        MysqlSideTableInfo mysqlTableInfo = new MysqlSideTableInfo();
-        mysqlTableInfo.setName(tableName);
-        parseFieldsInfo(fieldsInfo, mysqlTableInfo);
-
-        parseCacheProp(mysqlTableInfo, props);
-        mysqlTableInfo.setParallelism(MathUtil.getIntegerVal(props.get(MysqlSideTableInfo.PARALLELISM_KEY.toLowerCase())));
-        mysqlTableInfo.setUrl(MathUtil.getString(props.get(MysqlSideTableInfo.URL_KEY.toLowerCase())));
-        mysqlTableInfo.setTableName(MathUtil.getString(props.get(MysqlSideTableInfo.TABLE_NAME_KEY.toLowerCase())));
-        mysqlTableInfo.setUserName(MathUtil.getString(props.get(MysqlSideTableInfo.USER_NAME_KEY.toLowerCase())));
-        mysqlTableInfo.setPassword(MathUtil.getString(props.get(MysqlSideTableInfo.PASSWORD_KEY.toLowerCase())));
-
+        TableInfo mysqlTableInfo = super.getTableInfo(tableName, fieldsInfo, props);
+        mysqlTableInfo.setType(CURR_TYPE);
         return mysqlTableInfo;
     }
 }
