+++ mysql/mysql-side/pom.xml
@@ -11,13 +11,28 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>sql.side.mysql</artifactId>
+    <version>1.0-SNAPSHOT</version>
     <name>mysql-side</name>
+    <packaging>pom</packaging>
+
     <modules>
         <module>mysql-side-core</module>
         <module>mysql-async-side</module>
         <module>mysql-all-side</module>
     </modules>
 
-    <packaging>pom</packaging>
+
+    <properties>
+        <rdb.side.version>1.0-SNAPSHOT</rdb.side.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.dtstack.flink</groupId>
+            <artifactId>sql.side.rdb</artifactId>
+            <version>${rdb.side.version}</version>
+        </dependency>
+    </dependencies>
+
 
 </project>
