+++ rdb/rdb-sink/pom.xml
@@ -6,6 +6,7 @@
         <artifactId>sql.rdb</artifactId>
         <groupId>com.dtstack.flink</groupId>
         <version>1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
