+++ rdb/rdb-sink/src/main/java/com/dtstack/flink/sql/sink/rdb/table/RdbTableInfo.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -6,33 +6,30 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
- 
-
-package com.dtstack.flink.sql.sink.mysql.table;
+package com.dtstack.flink.sql.sink.rdb.table;
 
 import com.dtstack.flink.sql.table.TargetTableInfo;
 import org.apache.flink.calcite.shaded.com.google.common.base.Preconditions;
 
 /**
  * Reason:
- * Date: 2018/6/22
+ * Date: 2018/11/27
  * Company: www.dtstack.com
- * @author xuchao
+ *
+ * @author maqi
  */
+public class RdbTableInfo extends TargetTableInfo {
 
-public class MysqlTableInfo extends TargetTableInfo {
-
-    private static final String CURR_TYPE = "mysql";
+    //private static final String CURR_TYPE = "mysql";
 
     public static final String URL_KEY = "url";
 
@@ -47,10 +44,6 @@ public class MysqlTableInfo extends TargetTableInfo {
     public static final String BUFFER_SIZE_KEY = "bufferSize";
 
     public static final String FLUSH_INTERVALMS_KEY = "flushIntervalMs";
-
-    public MysqlTableInfo(){
-        setType(CURR_TYPE);
-    }
 
     private String url;
 
@@ -124,10 +117,10 @@ public class MysqlTableInfo extends TargetTableInfo {
 
     @Override
     public boolean check() {
-        Preconditions.checkNotNull(url, "mysql field of URL is required");
-        Preconditions.checkNotNull(tableName, "mysql field of tableName is required");
-        Preconditions.checkNotNull(userName, "mysql field of userName is required");
-        Preconditions.checkNotNull(password, "mysql field of password is required");
+        Preconditions.checkNotNull(url, "rdb field of URL is required");
+        Preconditions.checkNotNull(tableName, "rdb field of tableName is required");
+        Preconditions.checkNotNull(userName, "rdb field of userName is required");
+        Preconditions.checkNotNull(password, "rdb field of password is required");
         return true;
     }
