+++ sqlserver/pom.xml
@@ -6,30 +6,25 @@
         <artifactId>flink.sql</artifactId>
         <groupId>com.dtstack.flink</groupId>
         <version>1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>sql.sqlserver</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <modules>
-        <module>sqlserver-side</module>
-    </modules>
     <packaging>pom</packaging>
 
     <properties>
         <jtds.version>1.3.1</jtds.version>
         <sql.core.version>1.0-SNAPSHOT</sql.core.version>
-        <sql.side.rdb.version>1.0-SNAPSHOT</sql.side.rdb.version>
     </properties>
 
-    <dependencies>
-        <dependency>
-            <groupId>com.dtstack.flink</groupId>
-            <artifactId>sql.side.rdb</artifactId>
-            <version>${sql.side.rdb.version}</version>
-            <scope>provided</scope>
-        </dependency>
+    <modules>
+        <module>sqlserver-side</module>
+        <module>sqlserver-sink</module>
+    </modules>
 
+    <dependencies>
         <dependency>
             <groupId>com.dtstack.flink</groupId>
             <artifactId>sql.core</artifactId>
