+++ sqlserver/sqlserver-side/pom.xml
@@ -6,12 +6,12 @@
         <artifactId>sql.sqlserver</artifactId>
         <groupId>com.dtstack.flink</groupId>
         <version>1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>sql.side.sqlserver</artifactId>
     <version>1.0-SNAPSHOT</version>
-
     <name>sqlserver-side</name>
     <packaging>pom</packaging>
 
@@ -32,8 +32,6 @@
             <artifactId>sql.side.rdb</artifactId>
             <version>${rdb.side.version}</version>
         </dependency>
-
-
     </dependencies>
