+++ sqlserver/sqlserver-side/sqlserver-side-core/pom.xml
@@ -6,13 +6,14 @@
         <artifactId>sql.side.sqlserver</artifactId>
         <groupId>com.dtstack.flink</groupId>
         <version>1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>sql.side.sqlserver.core</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
-
+    <name>sqlserver-side-core</name>
 
 
 </project>
