+++ sqlserver/sqlserver-sink/pom.xml
@@ -1,22 +1,23 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>sql.mysql</artifactId>
+        <artifactId>sql.sqlserver</artifactId>
         <groupId>com.dtstack.flink</groupId>
         <version>1.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>sql.sink.mysql</artifactId>
-    <packaging>jar</packaging>
-
-    <name>mysql-sink</name>
-    <url>http://maven.apache.org</url>
+    <artifactId>sql.sink.sqlserver</artifactId>
+    <version>1.0-SNAPSHOT</version>
+    <name>sqlserver-sink</name>
 
     <properties>
         <sql.sink.rdb.version>1.0-SNAPSHOT</sql.sink.rdb.version>
     </properties>
+
 
     <dependencies>
         <dependency>
@@ -25,6 +26,7 @@
             <version>${sql.sink.rdb.version}</version>
         </dependency>
     </dependencies>
+
 
     <build>
         <plugins>
@@ -72,14 +74,14 @@
                         </goals>
                         <configuration>
                             <tasks>
-                                <copy todir="${basedir}/../../plugins/mysqlsink">
+                                <copy todir="${basedir}/../../plugins/sqlserversink">
                                     <fileset dir="target/">
                                         <include name="${project.artifactId}-${project.version}.jar"/>
                                     </fileset>
                                 </copy>
 
-                                <move file="${basedir}/../../plugins/mysqlsink/${project.artifactId}-${project.version}.jar"
-                                      tofile="${basedir}/../../plugins/mysqlsink/${project.name}.jar"/>
+                                <move file="${basedir}/../../plugins/sqlserversink/${project.artifactId}-${project.version}.jar"
+                                      tofile="${basedir}/../../plugins/sqlserversink/${project.name}.jar"/>
                             </tasks>
                         </configuration>
                     </execution>
