+++ sqlserver/sqlserver-sink/src/main/java/com/dtstack/flink/sql/sink/sqlserver/SqlserverSink.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -6,39 +6,42 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.dtstack.flink.sql.sink.sqlserver;
 
+import com.dtstack.flink.sql.sink.rdb.RdbSink;
 
-package com.dtstack.flink.sql.sink.mysql.table;
-
-import com.dtstack.flink.sql.sink.rdb.table.RdbSinkParser;
-import com.dtstack.flink.sql.table.TableInfo;
-
-import java.util.Map;
+import java.util.List;
 
 /**
  * Reason:
- * Date: 2018/7/4
+ * Date: 2018/11/27
  * Company: www.dtstack.com
  *
- * @author xuchao
+ * @author maqi
  */
-
-public class MysqlSinkParser extends RdbSinkParser {
-    private static final String CURR_TYPE = "mysql";
+public class SqlserverSink extends RdbSink {
+    private static final String SQLSERVER_DRIVER = "net.sourceforge.jtds.jdbc.Driver";
 
     @Override
-    public TableInfo getTableInfo(String tableName, String fieldsInfo, Map<String, Object> props) {
-        TableInfo mysqlTableInfo = super.getTableInfo(tableName, fieldsInfo, props);
-        mysqlTableInfo.setType(CURR_TYPE);
-        return mysqlTableInfo;
+    public String getDriverName() {
+        return SQLSERVER_DRIVER;
+    }
+
+    @Override
+    public void buildSql(String tableName, List<String> fields) {
+        buildInsertSql(tableName, fields);
+    }
+
+    private void buildInsertSql(String tableName, List<String> fields) {
+
     }
 }
