+++ sqlserver/sqlserver-sink/src/main/java/com/dtstack/flink/sql/sink/sqlserver/table/SqlserverSinkParser.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -6,18 +6,16 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-
-package com.dtstack.flink.sql.sink.mysql.table;
+package com.dtstack.flink.sql.sink.sqlserver.table;
 
 import com.dtstack.flink.sql.sink.rdb.table.RdbSinkParser;
 import com.dtstack.flink.sql.table.TableInfo;
@@ -26,19 +24,18 @@ import java.util.Map;
 
 /**
  * Reason:
- * Date: 2018/7/4
+ * Date: 2018/11/27
  * Company: www.dtstack.com
  *
- * @author xuchao
+ * @author maqi
  */
-
-public class MysqlSinkParser extends RdbSinkParser {
-    private static final String CURR_TYPE = "mysql";
+public class SqlserverSinkParser extends RdbSinkParser {
+    private static final String CURR_TYPE = "sqlserver";
 
     @Override
     public TableInfo getTableInfo(String tableName, String fieldsInfo, Map<String, Object> props) {
-        TableInfo mysqlTableInfo = super.getTableInfo(tableName, fieldsInfo, props);
-        mysqlTableInfo.setType(CURR_TYPE);
-        return mysqlTableInfo;
+        TableInfo sqlserverTableInfo = super.getTableInfo(tableName, fieldsInfo, props);
+        sqlserverTableInfo.setType(CURR_TYPE);
+        return sqlserverTableInfo;
     }
 }
